BigDecimal for money representation AbstractResponse extracted Deactivate accounts instead of removing

+++ src/main/java/pro/sholokhov/models/Transaction.java
@@ -2,20 +2,22 @@ package pro.sholokhov.models;
 
 import org.joda.time.DateTime;
 
+import java.math.BigDecimal;
+
 public class Transaction {
 
   private Long id;
   private DateTime date;
   private Long from;
   private Long to;
-  private Double amount;
+  private BigDecimal amount;
 
   public Transaction(Long id, DateTime date, Long from, Long to, Double amount) {
     this.id = id;
     this.date = date;
     this.from = from;
     this.to = to;
-    this.amount = amount;
+    this.amount = BigDecimal.valueOf(amount);
   }
 
 }
